Add pWebSocketServer
Adds pWebSocketServer, a program that translates MOOS Mail into WebSocket messages. To use, attach to the WebSocket at it's port (default 9090) and send a message with the variable you want to listen to (case sensitive). Whenever MOOS receives mail that a client subscribes to, it will send it in the format "VAR_NAME=value".

For now it uses messy code and external libraries that require C++11 and Boost, so tread with caution.

Signed-off-by: Conlan Cesar <conlanc@mit.edu>