Add password authentication
This adds a new mechanism for sending messages meant for pWebSocketServer itself. By prefixing messages with a '$' it signifies it as an internal message. Right now the only one is '$SetPassword' which, if enabled in the ProcessConfig, requires clients to send '$SetPassword=YOUR_PASSWORD' in order to be able to set further messages.