Store previous mail values to send them if new clients attach
Before, if a Client subscribed to a message, the message was sent, and another client connected that client would not see the most recent message. Therefore, store the most recent values of MOOS Mail in a simple map that is accessed when a new client initially subscribes to mail that already exists.

This is probably not the most efficient way to fix this, but there isn't much documentation on how to manually fetch the latest value of mail.