Add a new app 'iMumbleClient' with simple Mumble VoIP listening support
This app uses a Mumble C++ library to connect to a murmur server. As of right now, it will connect to a server at 'localhost' on the root channel and playback all audio. On the roadmap is transmitting and recording audio data via a mail trigger, and additional publishing of meta data.