Add recording and transmitting audio data to Mumble
This is a pretty big refactor compared to the last commit. It uses the Mumpi project as a reference for implementation, and switches from using the Blocking I/O Direct Write to using callbacks. This should lead to better performance, as it is the recommended method by PortAudio.

There is still work to be done for making the sent audio sound better, further MOOS logging and integration, and a channel system, but this is the first working duplex audio version.