Neptune: Add support for $MOEXT message
Updates the seebyte mission and iNeptune app to support the new $MOEXT NMEA message.

This can be used by Neptune to send encoded data to all other Neptune communities in the network. Message can be sent or recieved, following the same format in both cases.

Additionally, fixed the seebyte mission in general for better configuration, esp in regards to multi-vehicle deployments on the same machine.