Neptune: support for external vehicle driver comms
Right now to run a real vehicle, we need to run MOOS natively, or edit the docker image to contain the vehicle driver.
To resolve this, we forward the DESIRED_* mailings to an arbitrary community -- ideally only containing vehicle & sensor drivers. That community will recieve our DESIRED information, and send us back corresponding NAV_* and sensor messages.
This allows us to keep the Neptune image untouched, and run the container on any vehicle without major changes to the host or our image.