add pAnnounceMail app
Very very simple app -- takes any mailing configured, wraps it in a NODE_REPORT_LOCAL, and broadcasts it to all other vehicles.

Further enhancements could be done to appcasting or additional config vars, but I need this quickly for the Neptune mission.

Ideally this will be made redundant by enhancements to pEchoVar, but for now it is necessary. We'll see how long "for now" turns out to be :)